requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.5
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
auth0-python==4.7.2
av==13.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
cohere==5.13.3
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
crewai==0.86.0
crewai-tools==0.17.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.125
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.115.6
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googlesearch-python==1.3.0
gptcache==0.1.44
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
gspread==6.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ifaddr==0.2.0
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.7.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
json_repair==0.30.3
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.3.9
langchain-cohere==0.3.3
langchain-community==0.3.9
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langsmith==0.1.147
litellm==1.53.7
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mem0ai==0.1.34
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.8.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.57.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
peewee==3.17.9
pexpect==4.9.0
phi==0.6.7
phidata==2.7.10
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.4
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.10.1
pylance==0.20.0
pylibsrtp==0.11.0
pylint==3.3.1
pyOpenSSL==25.0.0
pyparsing==3.1.4
pypdf==5.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyright==1.1.390
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytube==15.0.0
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==26.1.0
qdrant-client==1.12.1
qrcode==7.4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
schema==0.7.7
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.3
setuptools==72.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.38.0
streamlit-webrtc==0.53.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uv==0.5.6
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
